Move `utils.rs` functions to `x509.rs` and remove module
