+++ src/prelude.rs
@@ -9,7 +9,6 @@ pub use crate::objects::*;
 pub use crate::pem::*;
 pub use crate::revocation_list::*;
 pub use crate::time::*;
-pub use crate::utils::*;
 #[cfg(feature = "validate")]
 pub use crate::validate::*;
 pub use crate::x509::*;
